@@ -7,9 +7,9 @@ namespace StravaDiscordBot
 {
     public class AppOptions
     {
+        public string ConnectionString { get; set; }
         public DiscordOptions Discord { get; set; }
         public StravaOptions Strava { get; set; }
-        public RemoteItOptions Remote { get; set; }
         public string StorageConnectionString { get; set; }
         public string BaseUrl { get; set; }
     }
@@ -22,14 +22,5 @@ namespace StravaDiscordBot
     {
         public string ClientId { get; set; }
         public string ClientSecret { get; set; }
-    }
-
-    public class RemoteItOptions
-    {
-        public string DeveloperKey { get; set; }
-        public string Username { get; set; }
-        public string Password { get; set; }
-        public string DeviceServiceId { get; set; }
-        public string DeviceName { get; set; }
     }
 }
