@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Azure.Cosmos.Table;
@@ -7,18 +8,21 @@ using StravaDiscordBot.Models.Strava;
 
 namespace StravaDiscordBot.Models
 {
-    public class LeaderboardParticipant : TableEntity
+    public class LeaderboardParticipant
     {
+        [Key]
+        public string DiscordUserId { get; set; }
+        public string ChannelId { get; set; }
         public string StravaAccessToken { get; set; }
         public string StravaRefreshToken { get; set; }
-        public string DiscordMention => $"<@{RowKey}>";
+        public string GetDiscordMention() => $"<@{DiscordUserId}>";
         public LeaderboardParticipant() {}
 
         // Don't like amount of the constructor parameters, but good enough for now
         public LeaderboardParticipant(string channelId, string userId, string stravaAccessToken, string stravaRefreshToken)
         {
-            PartitionKey = channelId;
-            RowKey = userId;
+            ChannelId = channelId;
+            DiscordUserId = userId;
             StravaAccessToken = stravaAccessToken;
             StravaRefreshToken = stravaRefreshToken;
         }
