@@ -11,12 +11,21 @@ namespace StravaDiscordBot
     {
         public static void Main(string[] args)
         {
-            Log.Logger = new LoggerConfiguration()
+            var logBuilder = new LoggerConfiguration()
             .MinimumLevel.Debug()
             .MinimumLevel.Override("Microsoft", LogEventLevel.Information)
             .Enrich.FromLogContext()
-            .WriteTo.Console()
-            .CreateLogger();
+            .WriteTo.Console();
+
+            Console.WriteLine($"Env: {Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT")}");
+
+            if (string.Equals(Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT"), "PRODUCTION", StringComparison.InvariantCultureIgnoreCase))
+            {
+                Console.WriteLine($"Log path: {Environment.GetEnvironmentVariable("LOG_PATH")}");
+                logBuilder.WriteTo.File(Environment.GetEnvironmentVariable("LOG_PATH"));
+            }
+
+            Log.Logger = logBuilder.CreateLogger();
 
             try
             {
