@@ -1,13 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Discord;
+﻿using System.Threading.Tasks;
 using Discord.WebSocket;
 using Microsoft.Extensions.Logging;
 using StravaDiscordBot.Exceptions;
-using StravaDiscordBot.Models;
-using StravaDiscordBot.Services.Storage;
 
 namespace StravaDiscordBot.Services.Discord.Commands
 {
@@ -34,11 +28,11 @@ namespace StravaDiscordBot.Services.Discord.Commands
 
             _logger.LogInformation($"Executing 'Join' command. Full: {message.Content} | Author: {message.Author}");
 
-            //if(await _stravaService.ParticipantAlreadyExistsAsync(message.Channel.Id.ToString(), message.Author.Id.ToString()))
-            //    throw new InvalidCommandArgumentException($"Whoops, it seems like you're already participating in the leaderboard");
+            if(await _stravaService.ParticipantAlreadyExistsAsync(message.Channel.Id.ToString(), message.Author.Id.ToString()))
+                throw new InvalidCommandArgumentException($"Whoops, it seems like you're already participating in the leaderboard");
 
             var dmChannel = await message.Author.GetOrCreateDMChannelAsync();
-            await dmChannel.SendMessageAsync($"Hey, {message.Author.Mention} ! Please go to this url to allow me check out your Strava activities: {await _stravaService.GetOAuthUrl(message.Channel.Id.ToString(), message.Author.Id.ToString())}");
+            await dmChannel.SendMessageAsync($"Hey, {message.Author.Mention} ! Please go to this url to allow me check out your Strava activities: {_stravaService.GetOAuthUrl(message.Channel.Id.ToString(), message.Author.Id.ToString())}");
         }
     }
 }
